requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
